Feat: migrar de PostgreSQL para MySQL e ajustar endpoints com /consulta
 - Realizado ajuste para o SQL Alchemy utilizar o MySQL Connector para buscar o banco MySQL do projeto; - Adicionado prefixo /consulta em todos os routers de leitura da API
