+++ backend/requirements.txt
@@ -1,7 +1,6 @@
 fastapi==0.111.0
 uvicorn==0.23.0
 sqlalchemy==2.0.23
-psycopg2-binary==2.9.7   # se usar PostgreSQL
-python-dotenv==1.0.0      # se quiser carregar variáveis de ambiente
+pymysql==1.1.0          # driver MySQL equivalente ao psycopg2
+python-dotenv==1.0.0
 uvicorn[standard]==0.23.2
-
