+++ backend/routers/consulta/categorias.py
@@ -3,7 +3,7 @@ from sqlalchemy import create_engine, text
 from sqlalchemy.orm import sessionmaker
 from config import load_config
 
-router = APIRouter()
+router = APIRouter(prefix="/consulta/categorias", tags=["Teste de Credenciais"]) 
 
 def create_db_session(env_type: str):
     DATABASE_URL, ENV_TYPE = load_config(env_type)
@@ -11,19 +11,18 @@ def create_db_session(env_type: str):
     SessionLocal = sessionmaker(bind=engine)
     return SessionLocal, ENV_TYPE
 
-@router.get("/tipopagamento")
+@router.get("/")
 
 
-def listar_pagamentos():
+def listar_categorias():
     SessionLocal, ENV_TYPE = create_db_session("leitura")
     db = SessionLocal()
     try:
-        result = db.execute(text("SELECT * FROM tipo_pagamento ORDER BY id")).fetchall()
+        result = db.execute(text("SELECT id, descricao FROM categoria ORDER BY descricao")).fetchall()
         return [{"id": row.id, #o nome definido nas aspas é o nome final que vai ser encontrado pelo JS, independentemente do nome da tabela.
-                 "descricao": row.descricao
-                 } for row in result]
+                 "descricao": row.descricao} for row in result]
     except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Erro ao buscar tipos de pagamento: {str(e)}")
+        raise HTTPException(status_code=500, detail=f"Erro ao buscar categorias: {str(e)}")
     finally:
         db.close()
